Add a rootDir option (#12)
This adds a `rootDir` option (`--rootDir` or `-r` on the command line) to specify the root directory of the project.  (It defaults to the current working directory.)  This also refactors options a bit: the options object is now stored so that it can be accessed globally via `getOptions()`, and when storing it paths are normalized and defaults filled in for missing values.