Improve path handling (#14)
This introduces the `NormalizedPath` type.  It's just an alias for `string`, but uses a trick to simulate nominal typing.  (See [Microsoft/Typescript #202](https://github.com/Microsoft/TypeScript/issues/202#issuecomment-302402671) for details.)  Any instance of `NormalizedPath` is guaranteed to be normalized because it can only be created by a call to `normalizePath`.  Any variables that contain paths throughout the code have been updated to use this type.

This also includes a minor fix to normalize the drive letter to upper-case in Windows paths.