Support restricting internal dependencies (#38)
With this change, you can add an `imports` configuration to `fence.json`.  If this config value is present, only imports that have one of the specified tags will be allowed.

```json
{
    "imports": [
        "tagA",
        "tagB"
    ]
}
```

Note: this only applies to code in the same project; to restrict external dependencies you can use the `dependencies` configuration.