Report error objects instead of just string messages (#46)
This is a breaking change, but should make error reporting more flexible in the future.  Now, instead of just reporting an error message string, we pass an object with various properties that may be useful to the consumer.  I've also revamped the error message to be a lot more detailed so that it's hopefully easier to track down errors when they happen.