Normalize config and options at load time (#47)
With this refactor, we now normalize the config objects at load time so that the rest of the app doesn't need to worry about all the possible variations in how the user may provide the config.  For the purposes of typing, `RawWhatever` refers to the type the user is allowed to provide, and `Whatever` is the normalized type that the rest of the app will consume.  E.g.

```
dependencies: ["foo", "bar"]
```

...gets normalized to...

```
dependencies: [
    {
        "dependency": "foo",
        "accessibleTo": null
    },
    {
        "dependency": "bar",
        "accessibleTo": null
    },
]
```

I also normalize the options object in the same way.