Change schema for export rules to be consistent with import and dependency rules (#61)
Previously the `export` field in a fence was a map of `(module glob) -> (set of tags it is accessible to)`.  This changes it to be similar to the `dependencies` field, an array of rules.  The rule can be a simple string (in which case it is considered exported with no restrictions) or it can be an object that defines the module glob and the set of tags it is accessible to.  See the README for details.

The old-style export schema is still supported in order make it easier to stage an upgrade, but it is deprecated and will be removed in an upcoming release.