Consider any import outside rootDir external (#76)
Using a Rush monorepo, imports from peer projects appear as coming from the local file system and not from node_modules. 

good-fences has a concept of a "rootDir" for the fence check. It seems by definition that anything that is not under that root should also be considered external for the sake of fencing.

This change modifies ImportRecord's isExternal check so that any import that is not under the rootDir is also considered external.