Introduce async SourceFileProvider interface
Adds a common interface that in future changes will be implemented
with a faster alternate that avoids scanning all files in the program
during initialization

As part of this work, I rework consumers of TypeScriptProgram to be
asyncronous, and add runWithConcurrentLimit to fence check in parallel.
This is needed because if we were to check all fences in parallel we
would hit the system RLIMIT by opening all fences simultaneously.

And while I'm in there, I added a progress bar to give user feedback
while performing fence checks. This does not display while performing
provider initialization, but this is less of a problem against the (yet
to come) newer implementation that doesn't perform full typescript
program initialization.