Allow comments in fence.json files (#110)
Add support for `//` style comments in fence.json files.

Support for this is inspired by eslint (which uses the same library to do it) and typescript (which seems to have its own implementation).

This allows easier documenting about specific dependencies, and why they are (or are not) important.
I have found these kinds of comments make maintaining/using fence.json files in project with many developers work better,
and it would be nice to be able to use `//` commends as this PR enables instead of adding comment fields to the JSON.

This works well with VS Code (and likely other tools) support for "JSON with comments".

Also included is a unit test for the BOM fix, since that is in the same function and didn't appear to have test coverage.

Like eslint, this uses version 3.1 of strip-json-comments. I tried version 4, but it uses ES modules which appear to break the jest configuration in this repo.